refactor(`JekyllSite`, `JekyllPage`): seperate `JekyllPage` from `JekyllSite`
